@@ -220,10 +220,10 @@ func InstallOrder(getter genericclioptions.RESTClientGetter, reg repo.IRegistry,
 				ChartSourceFlatRef: releasesapi.ChartSourceFlatRef{
 					Name:            pkg.Chart.Name,
 					Version:         pkg.Chart.Version,
-					SourceAPIGroup:  pkg.Chart.ChartRef.SourceRef.APIGroup,
-					SourceKind:      pkg.Chart.ChartRef.SourceRef.Kind,
-					SourceNamespace: pkg.Chart.ChartRef.SourceRef.Namespace,
-					SourceName:      pkg.Chart.ChartRef.SourceRef.Name,
+					SourceAPIGroup:  pkg.Chart.SourceRef.APIGroup,
+					SourceKind:      pkg.Chart.SourceRef.Kind,
+					SourceNamespace: pkg.Chart.SourceRef.Namespace,
+					SourceName:      pkg.Chart.SourceRef.Name,
 				},
 				Options: values.Options{
 					ValuesFile:  pkg.Chart.ValuesFile,
